requirements.txt
streamlit>=1.42.0
Authlib>=1.3.2
azure-appconfiguration>=1.4.0
azure-identity>=1.13.0
